Remove other author form readme and add module for addBook
