+++ README.md
@@ -18,14 +18,6 @@ This is a website for adding and removing books ddynamically using Javascript
 - Twitter: [@carpediem0011](https://twitter.com/carpediem0011)
 - LinkedIn: [LinkedIn](https://www.linkedin.com/in/david-zambrano-corral-b87a4198/)
 
-👤 **Abeera Tahir**
-
-- GitHub: [@githubhandle](https://github.com/AbeeraTahir)
-- Twitter: [@twitterhandle]( https://twitter.com/AbeeraTahir8?t=z5CjMpmHMZmS98i09gUpYA&s=08)
-- LinkedIn: [LinkedIn](https://www.linkedin.com/in/abeera-tahir-961893176)
-
-
-
 ## 🤝 Contributing
 
 Contributions, issues, and feature requests are welcome!
